fix: #111 web rtc error
