+++ lib/hooks/usePeerJS/usePeerJS.tsx
@@ -73,3 +73,8 @@ export function usePeerJS(options: { debug?: boolean }) {
 
   return { peer: peer.current, hostId, loading, error };
 }
+
+export function isWebRTCSupported() {
+  return false;
+  return RTCPeerConnection !== undefined;
+}
