+++ lib/routes/Play/JoinAGameRoute.tsx
@@ -17,6 +17,7 @@ import {
   CharactersContext,
   ICharacter,
 } from "../../contexts/CharactersContext/CharactersContext";
+import { isWebRTCSupported } from "../../hooks/usePeerJS/usePeerJS";
 import { useTranslate } from "../../hooks/useTranslate/useTranslate";
 
 let playerNameSingleton = "";
@@ -49,73 +50,93 @@ export const JoinAGame: React.FC<{
       <Box>
         <Box pb="1rem">
           <Container maxWidth="xs">
-            <form
-              onSubmit={(event) => {
-                event.preventDefault();
-                event.stopPropagation();
-                onSubmitPlayerName(playerName);
-              }}
-            >
-              <Box pb="2rem" textAlign="center">
-                <img width="150px" src={appIcon} />
-              </Box>
-              <Box pb="2rem" textAlign="center">
-                <Typography variant="h4">
-                  {t("play-route.connect-to-game")}
-                </Typography>
-              </Box>
-              <Box pb="1rem">
-                <InputLabel shrink>{t("play-route.character-name")}</InputLabel>
-                <TextField
-                  placeholder="Magnus Burnsides"
-                  value={playerName}
-                  onChange={(event) => {
-                    setPlayerName(event.target.value);
-                  }}
-                  inputProps={{
-                    maxLength: "50",
-                  }}
-                  fullWidth
-                  autoFocus
-                  required
-                />
-              </Box>
-              <Box pb="2rem">
-                <Grid container justify="center">
-                  <Grid item>
-                    <Button
-                      type="submit"
-                      variant={playerName ? "contained" : "outlined"}
-                      color="primary"
-                    >
-                      {playerName
-                        ? t("play-route.join-as", { playerName: playerName })
-                        : t("play-route.join")}
-                    </Button>
-                  </Grid>
-                </Grid>
-              </Box>
-              <Collapse in={props.connecting}>
-                <Box pb="1rem">
-                  <Box pb="3rem" display="flex" justifyContent="center">
-                    <Typography>{t("play-route.awesome-name")}</Typography>
-                  </Box>
-                  <Box display="flex" justifyContent="center">
-                    <CircularProgress />
-                  </Box>
-                </Box>
-              </Collapse>
-
-              <Collapse in={props.error}>
-                <Box pb="1rem" textAlign="center">
-                  <Typography color="error">
-                    {t("play-route.join-error")}
-                  </Typography>
-                </Box>
-              </Collapse>
-            </form>
+            {isWebRTCSupported() ? renderConnectionForm() : renderWebRTCError()}
           </Container>
         </Box>
+      </Box>
+    </Page>
+  );
+
+  function renderWebRTCError() {
+    return (
+      <Box>
+        <Box pb="2rem" textAlign="center">
+          <Typography variant="h4">{t("play-route.error.title")}</Typography>
+        </Box>
+        <Box pb="1rem" textAlign="center">
+          <Typography variant="body1">
+            {t("play-route.error.webRTC")}
+          </Typography>
+        </Box>
+      </Box>
+    );
+  }
+
+  function renderConnectionForm() {
+    return (
+      <form
+        onSubmit={(event) => {
+          event.preventDefault();
+          event.stopPropagation();
+          onSubmitPlayerName(playerName);
+        }}
+      >
+        <Box pb="2rem" textAlign="center">
+          <img width="150px" src={appIcon} />
+        </Box>
+        <Box pb="2rem" textAlign="center">
+          <Typography variant="h4">
+            {t("play-route.connect-to-game")}
+          </Typography>
+        </Box>
+        <Box pb="1rem">
+          <InputLabel shrink>{t("play-route.character-name")}</InputLabel>
+          <TextField
+            placeholder="Magnus Burnsides"
+            value={playerName}
+            onChange={(event) => {
+              setPlayerName(event.target.value);
+            }}
+            inputProps={{
+              maxLength: "50",
+            }}
+            fullWidth
+            autoFocus
+            required
+          />
+        </Box>
+        <Box pb="2rem">
+          <Grid container justify="center">
+            <Grid item>
+              <Button
+                type="submit"
+                variant={playerName ? "contained" : "outlined"}
+                color="primary"
+              >
+                {playerName
+                  ? t("play-route.join-as", { playerName: playerName })
+                  : t("play-route.join")}
+              </Button>
+            </Grid>
+          </Grid>
+        </Box>
+        <Collapse in={props.connecting}>
+          <Box pb="1rem">
+            <Box pb="3rem" display="flex" justifyContent="center">
+              <Typography>{t("play-route.awesome-name")}</Typography>
+            </Box>
+            <Box display="flex" justifyContent="center">
+              <CircularProgress />
+            </Box>
+          </Box>
+        </Collapse>
+
+        <Collapse in={props.error}>
+          <Box pb="1rem" textAlign="center">
+            <Typography color="error">{t("play-route.join-error")}</Typography>
+          </Box>
+        </Collapse>
+
         <Box>
           <Grid container justify="center">
             <Grid item>
@@ -133,9 +154,9 @@ export const JoinAGame: React.FC<{
             </Grid>
           </Grid>
         </Box>
-      </Box>
-    </Page>
-  );
+      </form>
+    );
+  }
 };
 
 JoinAGame.displayName = "JoinAGame";
