+++ lib/services/internationalization/locales/ptbrTranslations.ts
@@ -102,6 +102,7 @@ export const ptbrTranslation = {
   "play-route.error.description1": `Não pudemos conectá-lo(a) ao servidor para iniciar o jogo`,
   "play-route.error.description2": `Tente atualizar a página para ver se isso soluciona o problema. Você também pode iniciar um jogo offline.`,
   "play-route.error.title": `Algo errado aconteceu.`,
+  "play-route.error.webRTC": `O navegador que você está usando atualmente não é compatível com Fari para usar suas funcionalidades online. Escolha um navegador diferente, como Google Chrome ou Firefox.`,
   "play-route.fate-points": `Pontos de Destino`,
   "play-route.fp": `P.D.`,
   "play-route.init": `Iniciativa`,
